# regalloc: simplifies regInUseSet